Add final attempt + Custom Exceptions

+++ ClientCPP/ClientCPP/Common.h
@@ -11,7 +11,7 @@ typedef std::vector<uint8_t> Buffer;
 template<typename T, typename S>
 T convertTo(const S& s) {
 	T t{ 0 };
-	uint32_t amount = s.size() < t.size() ? s.size() : t.size(); // Take smaller
+	uint32_t amount = static_cast<uint32_t>(s.size() < t.size() ? s.size() : t.size()); // Take smaller
 	std::copy(s.cbegin(), s.cbegin() + amount, t.begin());
 	return t;
 }
